Add files via upload
Made a lotto machine that can create as many lines as you want (7numbers, 0-40), asks the user to save the line or print a new line. Displays the cost of the lines summed, and creates a .csv file that you can take to the kiosk and write down your numbers based on that file. And everytime the program is started it deletes the old line file.