Added SAS-app project
This is a java project that has a menu for the user and the user can choose from multiple options what he wants to proceed with.  There is an array defined to start with for options 1 & 2 (Linear and Binary sorting) but then the program will create a randomized array for options 3 & 4 (Bubble and Merge Sort). The program will run until user enters the Quit command.